Tor work
Resolved error where WebDAV would try to start Orbot, but still need to
patch up Tor proxy settings. Also wrapped Tor in a foreground service in
which I want to build out how we communicate status to the user. And
then leverage this architecture for the upload manager and "Snowbird" as
well.